requirements.txt
requests
kafka-python-ng
python-dotenv
psycopg2-binary
dotenv
google-genai
